# openapi.yaml
openapi: 3.0.3
info:
  title: Esketit Music Server API
  version: 1.3.0
  description: |
    REST API for listing and uploading media files, managing albums, tracks, authors,
    user playlists, and authenticating users with access and refresh tokens.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/album-covers:
    post:
      tags: [Album Covers]
      summary: Upload album cover
      description: |
        Upload an album cover file. The returned file can later be referenced from album
        `coverImagePath`. Preferred write format is `/api/album-covers/{name}`.

        The server preserves the original extension but rewrites the stored filename
        by appending a random suffix before the extension.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '201':
          description: Uploaded album cover metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/album-covers/{name}:
    get:
      tags: [Album Covers]
      summary: Get album cover file
      description: |
        Fetch an album cover file by filename. Album responses expose `coverImagePath`
        as a full absolute URL pointing to this route.
      parameters:
        - $ref: '#/components/parameters/FileName'
      responses:
        '200':
          description: Album cover file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/albums:
    get:
      tags: [Albums]
      summary: List albums
      description: |
        Returns albums for both admin and player clients. Admin requests can include
        albums with no tracks. Anonymous and listener requests omit albums whose
        `trackIds` array is empty.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: authorId
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: query
          in: query
          schema:
            type: string
        - name: isPublished
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Album page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAlbums'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Albums]
      summary: Create album
      description: Albums may be created as published even when `trackIds` is empty.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertAlbumRequest'
      responses:
        '201':
          description: Created album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/albums/{id}:
    get:
      tags: [Albums]
      summary: Get album by ID
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Albums]
      summary: Update album by ID
      description: Albums may remain published even when `trackIds` is empty.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertAlbumRequest'
      responses:
        '200':
          description: Updated album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Albums]
      summary: Delete album by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Album is used by one or more tracks
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PlainError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/search:
    get:
      tags: [Search]
      summary: Search authors, albums, and tracks
      description: |
        Public substring search across author names, album titles, and track names.
        Results are returned as a single combined paginated list.

        Anonymous and listener requests omit albums whose `trackIds` array is empty.
        Admin requests may include empty albums when authenticated.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: query
          schema:
            type: string
          description: Case-insensitive substring match
      responses:
        '200':
          description: Combined paginated search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSearchResults'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    FileName:
      name: name
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
    Forbidden:
      description: Insufficient permissions
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
...
    InternalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
  schemas:
    PlainError:
      type: string
    FileUploadRequest:
      type: object
      properties:
        file:
          type: string
          format: binary
      required: [file]
    FileInfo:
      type: object
      properties:
        name:
          type: string
        sizeBytes:
          type: integer
          format: int64
        lastModified:
          type: string
          format: date-time
        path:
          type: string
          description: Media path for the file, for example `/api/songs/example.mp3`.
        url:
          type: string
      required: [name, sizeBytes, lastModified, path, url]
...
    AdditionalInfo:
      type: object
      required: [type]
      properties:
        type:
          type: string
      additionalProperties: true
...
    Album:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        coverImagePath:
          type: string
          format: uri
          description: |
            Absolute URL to the album cover file. API responses normalize this field to a
            full URL pointing to `/api/album-covers/{name}`.
          example: http://192.168.1.6:8080/api/album-covers/%D0%93%D0%BE%D1%80%D0%B3%D0%BE%D1%80%D0%BE%D0%B4_2.jpg
        authorIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
        releaseDate:
          type: string
          format: date-time
        isPublished:
          type: boolean
        trackIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
          description: |
            May be empty. Non-admin album list responses omit albums whose `trackIds`
            array is empty.
        additionalInfo:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalInfo'
      required: [id, title, coverImagePath, authorIds, releaseDate, isPublished, trackIds, additionalInfo]
    UpsertAlbumRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        coverImagePath:
          type: string
          description: |
            Album cover reference. Preferred form is `/api/album-covers/{name}`.
            Absolute URLs are also accepted. Bare filenames may be normalized by the
            server, but clients should not rely on that form.
          example: /api/album-covers/%D0%93%D0%BE%D1%80%D0%B3%D0%BE%D1%80%D0%BE%D0%B4_2.jpg
        authorIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
          description: Ignored on write; server recalculates this field from album tracks.
        releaseDate:
          type: string
          format: date-time
        isPublished:
          type: boolean
        trackIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
          description: May be empty when creating or updating an album.
        additionalInfo:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalInfo'
      required: [title, releaseDate, isPublished, trackIds, additionalInfo]
...
    PaginatedAlbums:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
      required: [items, page, pageSize, totalItems, totalPages]
...
    SearchResultItem:
      type: object
      properties:
        type:
          type: string
          enum: [author, album, track]
        author:
          $ref: '#/components/schemas/Author'
        album:
          $ref: '#/components/schemas/Album'
        track:
          $ref: '#/components/schemas/Track'
      required: [type]
    PaginatedSearchResults:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultItem'
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
      required: [items, page, pageSize, totalItems, totalPages]
...